tags in objects reflected post instantiation changes to classes and roles
The lazy building of the tags meant that if classes or roles were
changed after an object was instantiated, the tags for an object reflected
those changes.

The "fix" is to freeze the tags for an object at instantiation.  This
however means that any tag changes made to the object after
instantiation are ignored.