test for additions of tags to instantiated class were broken.
By checking for an object's tags prior to applying a role to the
class, the object performed an explicit cache of the existing tags,
subverting any later change to the class.

So, don't call _tags prior to applying the role.  The object should
be caching at instantiation, not when _tags is called.